requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
openai==1.3.0
httpx==0.25.0
langchain>=0.2.0
langchain-core>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
langchain-huggingface>=0.0.1  # Для HuggingFaceEmbeddings в новых версиях LangChain
pypdf>=6.0.0  # Для PyPDFLoader в langchain-community (автоматически устанавливается, но лучше указать явно)
chromadb>=0.4.18  # Для fallback, может быть несовместим с pydantic 2.x
qdrant-client>=1.16.0  # Обновлено для поддержки Python 3.14
sentence-transformers==2.2.2
python-dotenv==1.0.0
aiofiles==23.2.1
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
PyMuPDF==1.23.8  # Для конвертации PDF в изображения для Vision API
Pillow==10.1.0  # Для работы с изображениями
tenacity==8.2.3
pybreaker==1.0.2
loguru==0.7.2
celery==5.3.4
redis==5.0.1
flower==2.0.1
mlflow==2.8.1

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.0

